psalm/model/mask_decoder/Mask2Former_Simplify/requirements.txt
addict==2.4.0
detectron2==0.6+cu111
fvcore==0.1.5.post20220512
imgaug==0.4.0
matplotlib==3.6.1
mmcv==1.6.2
MultiScaleDeformableAttention==1.0
numpy==1.17.4
opencv_python==4.6.0.66
Pillow==9.3.0
scipy==1.3.3
setuptools==45.2.0
SimpleITK==2.2.0
tensorboardX==2.5.1
timm==0.6.11
torch==1.10.0+cu111
torchvision==0.11.0+cu111
tqdm==4.64.1
